+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for setting a timeout duration to test.
 - Added option to CLI for generating code with custom words.
 - Added an function to generate code from `&str` (words split by comma).
+- Added option to CLI for printing results in JSON format.
 
 ### Changed
 - The way of calculating consistency is changed for better overview. Now it continues even if an error occures.